+++ example/lib/home/widgets/scan_filter_widget.dart
@@ -60,8 +60,7 @@ class _ScanFilterWidgetState extends State<ScanFilterWidget> {
             throw Exception("Invalid Manufacturer Data $manufacturer");
           }
           manufacturerDataFilters.add(
-            ManufacturerDataFilter(companyIdentifier: companyIdentifier),
-          );
+              ManufacturerDataFilter(companyIdentifier: companyIdentifier));
         }
       }
 
@@ -115,19 +114,19 @@ class _ScanFilterWidgetState extends State<ScanFilterWidget> {
           const Divider(),
           const SizedBox(height: 10),
           TextFormField(
-            controller: widget.servicesFilterController,
+            controller: widget.namePrefixController,
             maxLines: 2,
             decoration: const InputDecoration(
-              labelText: "Services",
+              labelText: "Name Prefixes",
               border: OutlineInputBorder(),
             ),
           ),
           const SizedBox(height: 10),
           TextFormField(
-            controller: widget.namePrefixController,
+            controller: widget.servicesFilterController,
             maxLines: 2,
             decoration: const InputDecoration(
-              labelText: "Name Prefixes",
+              labelText: "Services",
               border: OutlineInputBorder(),
             ),
           ),