+++ lib/src/universal_ble_linux/universal_ble_linux.dart
@@ -107,12 +107,12 @@ class UniversalBleLinux extends UniversalBlePlatform {
       _deviceAdded = null;
       _deviceRemoved = null;
 
-      // Stop Disovery
+      // Stop Discovery
       if (_activeAdapter?.discovering == true) {
         await _activeAdapter?.stopDiscovery();
       }
 
-      // Clean all advertiseemnt listeners
+      // Clean all advertisement listeners
       _deviceAdvertisementSubscriptions.removeWhere((e, value) {
         value.cancel();
         return true;
@@ -350,6 +350,7 @@ class UniversalBleLinux extends UniversalBlePlatform {
   Future<List<BleDevice>> getSystemDevices(
     List<String>? withServices,
   ) async {
+    await _ensureInitialized();
     List<BlueZDevice> devices =
         _client.devices.where((device) => device.connected).toList();
     if (withServices != null && withServices.isNotEmpty) {
@@ -455,7 +456,7 @@ class UniversalBleLinux extends UniversalBlePlatform {
 
   void _onDeviceAdd(BlueZDevice device) {
     BleDevice bleDevice = device.toBleDevice();
-    if (!_bleFilter.filterDevice(bleDevice)) {
+    if (!_bleFilter.matchesDevice(bleDevice)) {
       return;
     }
 
@@ -475,7 +476,7 @@ class UniversalBleLinux extends UniversalBlePlatform {
             e.contains(BluezProperty.manufacturerData) ||
             e.contains(BluezProperty.uuids))
         .listen((_) {
-      if (_bleFilter.filterDevice(bleDevice)) {
+      if (_bleFilter.matchesDevice(bleDevice)) {
         updateScanResult(device.toBleDevice());
       }
     });